+++ packages/app/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @openspecui/app
 
+## 6.0.1
+
 ## 6.0.0
 
 ### Major Changes